// apps/admin-web/src/.generated/api/model/operatorStatus.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Admin API
 * Control Plane Admin API Service
 * OpenAPI spec version: 1.0.0
 */

export type OperatorStatus = typeof OperatorStatus[keyof typeof OperatorStatus];


export const OperatorStatus = {
  active: 'active',
  banned: 'banned',
  deleted: 'deleted',
} as const;